Refactoring: generification.
The goal of this change is to make search by game tree to
evaluate moves independent from our actual implementation classes (Position/Move)
thus the algorithm should become universal.

Evaluator interface is implemented in multiple places so to reduce
amount of changes I extracted a generic parent.

Brain interface is implemented just in 3 places so I decided to make
it generic in-place.